[ISSUE-27] Refactored FeatureEmoji component
