+++ src/Components/Commom/FeatureEmoji/FeatureEmoji.spec.tsx
@@ -1,15 +1,15 @@
 import {render} from "@testing-library/react";
 import {Provider} from "react-redux";
+import Store from "../../../Redux/Store";
 import {BrowserRouter} from "react-router-dom";
 import React from "react";
-import Store from "../../../Redux/Store";
-import FacebookButton from "./index";
+import FeatureEmoji from "./index";
 
 test('matches snapshot', () => {
     const component = render(
         <Provider store={Store.store}>
             <BrowserRouter>
-                <FacebookButton onClick={jest.fn()}/>
+                <FeatureEmoji name="any"/>
             </BrowserRouter>
         </Provider>
     );
